requirements.txt
torch
torchvision
pyyaml
tqdm
scipy
tensorboard
prettytable
munch